Cleaned up non focused pipeline
